requirements.txt
# FastAPI and ASGI Server
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-multipart==0.0.12

# Database - SQLAlchemy with async support
sqlalchemy==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.3

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==43.0.1
python-multipart==0.0.12

# Instagram Integration
instaloader==4.13
instagrapi==2.1.2

# Settings and Configuration
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# HTTP Client
httpx==0.27.2

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0

# Code Quality
black==24.8.0
ruff==0.6.9
mypy==1.11.2

# Utilities
python-dateutil==2.9.0
